downgrades the to python 3.11 for the debugger
the pycharm debugger was giving me trouble with 3.12, so I downgraded; I
didn't investigate it very much, so there could be an easy fix.
The error was in the DANDI streaming code with a context manager or
generator not having a method, which caused the debugger to hang.
But pytest worked, so I think it's pycharm.